Divide the void-storage session impl and repoint its ownership baseline (#707)
`handlers/void_storage.rs` was 1,681 lines because one `impl WorldSession`
held 23 methods and 1,408 of those lines:

| group | methods | lines |
| --- | --- | --- |
| items | 10 | 228 |
| publication | 7 | 155 |
| unlock_and_query | 3 | 121 |
| transfer | 2 | 881 |
| swap | 1 | 95 |

The root keeps the free functions, the `Default` impl and the mounts at 244
lines. `handle_void_storage_transfer_with_generators_like_cpp` is 865 lines by
itself and stays whole: cutting one C++-mirroring operation into pieces is a
behaviour question, not a physical one, the same conclusion #621 recorded for
session/spell_effects/execution.rs.

The source-contract test needed a real change, not a path fix. It read
void_storage.rs with `include_str!` and sliced three handler bodies out of it
by their signatures to assert no `CharStatements`, `SqlTransaction`,
`.prepare(` or `char_db` appears there. Those signatures no longer live in
that file, so the slice returned nothing and the test panicked. It now audits
every file of the divided module - the root and all five submodules - which
is strictly more coverage than three handler bodies, and asserts that the set
of audited files equals the `.rs` files actually present in
`handlers/void_storage/`, so a submodule added later cannot escape the audit
by not being listed.

The ownership baseline is repointed rather than regenerated blindly. The
delta was produced with `session-ownership-check print-baseline` and diffed
field by field against the recorded baseline:

- `impls`: the single `crate::handlers::void_storage` entry becomes the five
  submodule paths. Ignoring the module field, the entries are identical.
- `impl_items`: 3,684 entries before and after. The 23 void-storage methods
  change module, and 18 of them gain the `visible` guard because a sibling
  submodule now calls them - the reach they already had inside one file.
- `registry_accesses` (594) and `bridge_accesses` (65) keep their counts; the
  three changed rows are #697's `player_quest_persistence_projection` and
  `player_directory_canonical_queries` renames, which that delivery left
  stale in this baseline.
- `persistence_access_snapshot` is byte-identical, so no persistence access
  moved.

Method multiset identical. wow-world's 3,822 tests pass, the 33 registry
metadata tests pass, no target gains a warning, and architecture check,
self-test, physical ratchet, hotspot ratchet and the syntax-only session
ownership check pass - the last now reporting 66 impl owners for the same
3,684 items.

Closes #707
